String serialization robustly handles encoding
When serializing string data, first we try encoding with utf-8. If that
raises an encoding error, we then try forcing the encoding to utf-8 and
see if it's valid before raising the exception.

This protects against cases where, e.g., data from a remote source is
returned as binary data, but is in fact utf-8 data.