Enable configurable retry periods on primary ops.
Moped::Session now can take 2 additional options, :retry_interval which
is the time to wait before attempting an operation on a primary again,
and :max_retries which is the maximum number of times to attempt the
operation before raising an error.

:retry_interval defaults to 1 (second).
:max_retries defaults to 30.

Instead of Moped attempting to perform an operation on the primary twice
before raising an error, it now refreshes the node list and retries
again up to the max number at retry intervals.

Related to: #80, #60

Thanks to Nicolas Viennot for the initial pull request to set the base
for the changes.