Hooking in new read preferences and write concerns.
The use of consistency: :strong/:eventual for reads is now completely
removed. The new syntax is in line with MongoDB's read preferences as
outlined here:

http://docs.mongodb.org/manual/applications/replication/#read-preference

In Moped, this is:

read: :primary
read: :primary_preferred
read: :secondary
read: :secondary_preferred
read: :nearest

For write concerns, the :safe keyword is no longer used, we now use
:write and the default for all writes is to propagate (ie safe mode on).

In Moped, the new syntax is:

write: :propagate
write: :unverified
write: { w: 3 }
write: { fsync: true }

etc...

The default max retries for failover is now equal to the number of nodes
in the cluster, instead of the purely random number 20.

[ close #92 ]
[ #133 ]