Query cursor should always return a Enumarable and not Enumerator.
The Enumerator.next implementation is different on MRI and JRuby.
JRuby spins a new thread to fetch the .next element,
which will cause Moped to get a new connection in the pool.
Having those different implementations, lead us to not return Enumerator to avoid user mistakes.

For an example of the different implementation see: https://gist.github.com/arthurnn/6734205