Ensure mongodb credentials are always sent upon (re)connection
If existing socket is reset, perhaps due to mongod failure,
credentials hash isn't reset. This causes them to not be sent upon
reconnection. Changed to reset at connect() instead of relying on a
disconnect() that may or may not happen.

Also cache last set of credentials so they can be reused when connection
loss is detected by with_connection().